pyproject.toml
[project]
name = "sinasc-research"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dash>=3.2.0",
    "dash-bootstrap-components>=2.0.4",
    "geoalchemy2>=0.18.0",
    "geopandas>=1.1.1",
    "gunicorn>=23.0.0",
    "pandarallel>=1.6.5",
    "pandas>=2.3.3",
    "plotly>=6.3.0",
    "psycopg2>=2.9.10",
    "pyarrow>=21.0.0",
    "pygrowthstandards>=0.1.2",
    "python-dotenv>=1.1.1",
    "seaborn>=0.13.2",
    "shapely>=2.1.2",
    "sqlalchemy>=2.0.43",
    "tqdm>=4.67.1",
]

[dependency-groups]
dev = [
    "ruff>=0.13.3",
]
notebok = [
    "ipywidgets>=8.1.7",
    "notebook>=7.4.7",
]

[tool.ruff]
line-length = 144
